+++ changelog.md
@@ -1,4 +1,17 @@
 # Changelog
+### v1.0.3 - 27-07-2021
+
+Additions:
+<ul>
+<li>Added support for controlling Soundscape through other modules or macros</li>
+<li>Added (potential) support for Playlist Enhancer</li>
+</ul>
+
+Fixes:
+<ul>
+<li>Fixed the import function</li>
+</ul>
+
 ### v1.0.2 - 14-07-2021
 -Playlists and playlist sounds can not be dragged into the mixer to populare the soundboard or a channel
 -Forgot to fix capitalization issues for the soundboard config, which prevented it from opening